backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
gTTS==2.5.0
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
